# generated/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ConfigData": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CriteriaSet": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"criterias": {
						"items": {
							"$ref": "#/components/schemas/Criteria"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"criterias"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Goal": {
				"properties": {
					"title": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"completed": {
						"type": "number",
						"format": "double"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/Goal"
						},
						"type": "array"
					},
					"parent": {
						"$ref": "#/components/schemas/Goal"
					},
					"owner": {
						"$ref": "#/components/schemas/GoalSummery"
					},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"goalType": {
						"$ref": "#/components/schemas/GoalType"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"teamWeight": {
						"type": "number",
						"format": "double"
					},
					"currentValue": {
						"type": "number",
						"format": "double"
					},
					"goalMark": {
						"$ref": "#/components/schemas/GoalMark"
					},
					"progress": {
						"$ref": "#/components/schemas/Progress"
					},
					"reusable": {
						"type": "boolean"
					}
				},
				"required": [
					"title",
					"id",
					"description",
					"status",
					"children",
					"parent",
					"owner",
					"goalType",
					"startDate",
					"endDate",
					"goalMark",
					"progress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoalSummery": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"cycle": {
						"$ref": "#/components/schemas/PerformanceCycle"
					},
					"cycleId": {
						"type": "number",
						"format": "double"
					},
					"owner": {
						"$ref": "#/components/schemas/GoalHolder"
					},
					"goalsAvailability": {
						"type": "boolean"
					},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"goals": {
						"items": {
							"$ref": "#/components/schemas/Goal"
						},
						"type": "array"
					},
					"approvalStatus": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"completed": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"duration": {
						"items": {
							"$ref": "#/components/schemas/TimeDuration"
						},
						"type": "array"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"teamId": {
						"type": "number",
						"format": "double"
					},
					"teamLeader": {
						"type": "boolean"
					},
					"peerReview": {
						"items": {
							"$ref": "#/components/schemas/PeerReview"
						},
						"type": "array"
					},
					"subordinateReviews": {
						"items": {
							"$ref": "#/components/schemas/SubordinateReview"
						},
						"type": "array"
					},
					"criteriaSet": {
						"$ref": "#/components/schemas/CriteriaSet"
					},
					"criteriaSetId": {
						"type": "number",
						"format": "double"
					},
					"configurations": {
						"items": {
							"$ref": "#/components/schemas/TeamConfiguration"
						},
						"type": "array"
					},
					"leadConfigurations": {
						"items": {
							"$ref": "#/components/schemas/TeamConfiguration"
						},
						"type": "array"
					},
					"internalCustomers": {
						"items": {
							"$ref": "#/components/schemas/InternalCustomer"
						},
						"type": "array"
					},
					"comment": {
						"$ref": "#/components/schemas/SupervisorAnalysis"
					},
					"internalCustomerApprover": {
						"type": "number",
						"format": "double"
					},
					"publishForPeers": {
						"type": "boolean"
					},
					"publishForLeaderPeers": {
						"type": "boolean"
					},
					"publishForInternalCustomers": {
						"type": "boolean"
					},
					"cycleClose": {
						"type": "boolean"
					},
					"finalCommentSubmit": {
						"type": "boolean"
					},
					"approvedInternalCustomers": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerMarks": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerLeaderMarks": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerCount": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerWeight": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerLeaderWeight": {
						"type": "number",
						"format": "double"
					},
					"approvedPeerReview": {
						"type": "number",
						"format": "double"
					},
					"peerReviewMarks": {
						"type": "number",
						"format": "double"
					},
					"peerReviewCount": {
						"type": "number",
						"format": "double"
					},
					"peerReviewWeight": {
						"type": "number",
						"format": "double"
					},
					"teamPerformanceWeight": {
						"type": "number",
						"format": "double"
					},
					"approvedTeamPerformance": {
						"type": "number",
						"format": "double"
					},
					"teamPerformanceCount": {
						"type": "number",
						"format": "double"
					},
					"teamPerformanceMarks": {
						"type": "number",
						"format": "double"
					},
					"supervisorFeedBackWeight": {
						"type": "number",
						"format": "double"
					},
					"approvedSupervisorFeedBack": {
						"type": "number",
						"format": "double"
					},
					"supervisorFeedBackCount": {
						"type": "number",
						"format": "double"
					},
					"supervisorFeedBackMarks": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"cycle",
					"owner",
					"goalsAvailability",
					"goals",
					"duration",
					"team",
					"peerReview",
					"subordinateReviews",
					"criteriaSet",
					"configurations",
					"leadConfigurations",
					"internalCustomers",
					"comment",
					"cycleClose",
					"finalCommentSubmit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PerformanceCycle": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"goalSummaries": {
						"items": {
							"$ref": "#/components/schemas/GoalSummery"
						},
						"type": "array"
					},
					"cycleNotifications": {
						"items": {
							"$ref": "#/components/schemas/CycleNotification"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"startDate",
					"endDate",
					"status",
					"goalSummaries",
					"cycleNotifications"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Team": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"teamType": {
						"$ref": "#/components/schemas/TeamType"
					},
					"teamTypeId": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"leads": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"parent": {
						"$ref": "#/components/schemas/Team"
					},
					"parentId": {
						"type": "number",
						"format": "double"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"userGroupMembers": {
						"items": {
							"$ref": "#/components/schemas/UserGroupMember"
						},
						"type": "array"
					},
					"userGroupMembers1": {
						"items": {
							"$ref": "#/components/schemas/UserGroupMember"
						},
						"type": "array"
					},
					"image": {
						"type": "string"
					},
					"goalHolder": {
						"$ref": "#/components/schemas/GoalHolder"
					},
					"goalHolderId": {
						"type": "number",
						"format": "double"
					},
					"memberCount": {
						"type": "number",
						"format": "double"
					},
					"cycle": {
						"$ref": "#/components/schemas/PerformanceCycle"
					},
					"cycleId": {
						"type": "number",
						"format": "double"
					},
					"previous": {
						"$ref": "#/components/schemas/Team"
					},
					"previousId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"teamType",
					"leads",
					"teams",
					"children",
					"members",
					"goalHolder",
					"memberCount",
					"cycle",
					"previous"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TeamType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"color",
					"teams"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"employeeId": {
						"type": "string"
					},
					"epfNo": {
						"type": "number",
						"format": "double"
					},
					"designation": {
						"$ref": "#/components/schemas/Designation"
					},
					"supervisor": {
						"$ref": "#/components/schemas/User"
					},
					"email": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"doj": {
						"type": "string"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"password": {
						"items": {
							"$ref": "#/components/schemas/Password"
						},
						"type": "array"
					},
					"goalHolder": {
						"$ref": "#/components/schemas/GoalHolder"
					},
					"goalsAvailability": {
						"type": "boolean"
					},
					"goalHolderId": {
						"type": "number",
						"format": "double"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"teams1": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"userGroupMembers": {
						"items": {
							"$ref": "#/components/schemas/UserGroupMember"
						},
						"type": "array"
					},
					"notifications": {
						"items": {
							"$ref": "#/components/schemas/Notification"
						},
						"type": "array"
					},
					"delete": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"name",
					"employeeId",
					"epfNo",
					"designation",
					"email",
					"password",
					"goalHolder",
					"goalsAvailability",
					"teams",
					"teams1",
					"notifications",
					"delete"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Designation": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					},
					"user": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"value",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Password": {
				"properties": {
					"password": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"password",
					"id",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoalHolder": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"goalsAvailability": {
						"type": "boolean"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"teamId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"goalSummaries": {
						"items": {
							"$ref": "#/components/schemas/GoalSummery"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"goalsAvailability",
					"goalSummaries"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserGroupMember": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"teamId": {
						"type": "number",
						"format": "double"
					},
					"teamLeads": {
						"$ref": "#/components/schemas/Team"
					},
					"teamLeadsId": {
						"type": "number",
						"format": "double"
					},
					"userGroup": {
						"$ref": "#/components/schemas/UserGroup"
					},
					"userGroupId": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string",
						"enum": [
							"user",
							"teamMember",
							"teamLead"
						]
					}
				},
				"required": [
					"id",
					"user",
					"team",
					"teamLeads",
					"userGroup",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserGroup": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"userGroupMembers": {
						"items": {
							"$ref": "#/components/schemas/UserGroupMember"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"userGroupMembers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Notification": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"time": {
						"type": "string",
						"format": "date-time"
					},
					"url": {
						"type": "string"
					},
					"read": {
						"type": "boolean"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"image",
					"time",
					"url",
					"read",
					"userId",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimeDuration": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"goalSummery": {
						"$ref": "#/components/schemas/GoalSummery"
					},
					"goalSummeryId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"from",
					"to",
					"goalSummery"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PeerReview": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"type": "number",
						"format": "double"
					},
					"userGoalSummery": {
						"$ref": "#/components/schemas/GoalSummery"
					},
					"userGoalSummeryId": {
						"type": "number",
						"format": "double"
					},
					"reviewer": {
						"type": "number",
						"format": "double"
					},
					"leader": {
						"type": "boolean"
					},
					"goalSummery": {
						"$ref": "#/components/schemas/GoalSummery"
					},
					"goalSummeryId": {
						"type": "number",
						"format": "double"
					},
					"peerReviewAnswer": {
						"items": {
							"$ref": "#/components/schemas/PeerReviewAnswer"
						},
						"type": "array"
					},
					"peerReviewHistory": {
						"items": {
							"$ref": "#/components/schemas/PeerReviewHistory"
						},
						"type": "array"
					},
					"overallComment": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"user",
					"userGoalSummery",
					"reviewer",
					"goalSummery",
					"peerReviewAnswer",
					"peerReviewHistory",
					"overallComment",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SubordinateReview": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"type": "number",
						"format": "double"
					},
					"userGoalSummery": {
						"$ref": "#/components/schemas/GoalSummery"
					},
					"userGoalSummeryId": {
						"type": "number",
						"format": "double"
					},
					"goalSummery": {
						"$ref": "#/components/schemas/GoalSummery"
					},
					"goalSummeryId": {
						"type": "number",
						"format": "double"
					},
					"reviewer": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"reviewType": {
						"type": "string"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"subordinateReviewAnswer": {
						"items": {
							"$ref": "#/components/schemas/SubordinateReviewAnswer"
						},
						"type": "array"
					},
					"subordinateReviewHistory": {
						"items": {
							"$ref": "#/components/schemas/SubordinateReviewHistory"
						},
						"type": "array"
					},
					"overallComment": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"subordinateReviewCriteriaSet": {
						"$ref": "#/components/schemas/CriteriaSet"
					},
					"subordinateReviewCriteriaSetId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"user",
					"userGoalSummery",
					"goalSummery",
					"reviewer",
					"type",
					"reviewType",
					"weight",
					"subordinateReviewAnswer",
					"subordinateReviewHistory",
					"overallComment",
					"status",
					"subordinateReviewCriteriaSet"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GoalType": {
				"type": "string",
				"enum": [
					"objective",
					"keyResult"
				]
			},
...
			"NotificationData": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"time": {
						"type": "string",
						"format": "date-time"
					},
					"url": {
						"type": "string"
					},
					"read": {
						"type": "boolean"
					},
					"user": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"description",
					"url",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PerformanceCycleData": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"notification": {}
				},
				"required": [
					"name",
					"startDate",
					"endDate",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DataTableResponse_User_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"description": "Bearer authentication - give access to the bearer of this token",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "pms-backend",
		"description": "Performance Management System Backend",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Tharindu vindula"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/config": {
			"get": {
				"operationId": "GetConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConfigData"
							}
						}
					}
				}
			}
		},
		"/api/v1/config/{id}": {
			"patch": {
				"operationId": "EditConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConfigData"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/notification": {
			"post": {
				"operationId": "CreateNotification",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NotificationData"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetNotification",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "date",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/notification/{userId}": {
			"patch": {
				"operationId": "MarkAsRead",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"all": {
										"type": "boolean"
									},
									"notificationId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"notificationId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/cycles": {
			"get": {
				"operationId": "GetCycle",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"manager",
							"hr"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddCycle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PerformanceCycleData"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/cycles/{id}": {
			"patch": {
				"operationId": "EditCycle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PerformanceCycleData"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCycle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/users/all/performance/{cycleId}": {
			"get": {
				"operationId": "GetUsersPerformance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DataTableResponse_User_"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cycleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "email",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "employeeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "epfNumber",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "exportValue",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}